Squashed commit of the following:
commit 57b781e5142835be26776efe32bc74188850a20e
Merge: 79abb32 71799a8
Author: Bum <92982989+blue7111@users.noreply.github.com>
Date:   Tue Sep 5 15:20:23 2023 +0900

    Merge pull request #68 from codestates-seb/feat/jongbum

    [Fix] axios inter

commit 71799a88150f8478d71a68fcacd51139da354c26
Author: LeeJongBum <qja965@gmail.com>
Date:   Tue Sep 5 15:17:57 2023 +0900

    [Fix] axios inter]

commit 79abb32a441ee65b28d385c0084f5e951d016034
Merge: d1d0e37 5833f2b
Author: Bum <92982989+blue7111@users.noreply.github.com>
Date:   Tue Sep 5 14:28:03 2023 +0900

    Merge pull request #67 from codestates-seb/feat/jongbum

    [Feat] Profile Valid

commit 5833f2bf2b0b3ea3ccccb9c8202037c6415db17c
Merge: 5071292 d1d0e37
Author: Bum <92982989+blue7111@users.noreply.github.com>
Date:   Tue Sep 5 14:27:53 2023 +0900

    Merge branch 'clientDev' into feat/jongbum

commit 5071292cd728620c4f60ac3d7be4f54d6d60829e
Author: LeeJongBum <qja965@gmail.com>
Date:   Tue Sep 5 14:20:34 2023 +0900

    [Feat] Profile Valid

commit d1d0e375f34dd1ab1e6f31c0378b844b9e8696bb
Merge: f986b39 9be97a7
Author: 박찬섭 <117289923+qkrckstjq@users.noreply.github.com>
Date:   Tue Sep 5 09:19:38 2023 +0900

    Merge pull request #66 from codestates-seb/feat/chansub

    [Design] ProGress 추가

+++ client/src/components/profile/ProfileCard.jsx
@@ -4,17 +4,23 @@ import { PiUploadSimple } from 'react-icons/pi';
 import useNav from '../../hooks/useNav';
 import { useParams } from 'react-router-dom';
 import api from '../../hooks/useAxiosInterceptor';
-import { deleteUser } from '../../redux/userform/userslice';
-import { useDispatch } from 'react-redux';
+import { updateUser, deleteUser } from '../../redux/userform/userslice';
+import { useSelector, useDispatch } from 'react-redux';
 import EditPassword from './EditPassword';
 import EditProfile from './EditProfile';
 import Withdrawal from './Withdrawal';
 import ShowProfile from './ShowProfile';
+import { isValidPassword } from './isValid';
+import userDefaultImg from '../../static/images/userDefaultImg.jpeg';
 
 const StyleProfileContainer = styled.div`
   display: flex;
+  padding-top: 3rem;
   gap: 5rem;
   font-size: 2rem;
+  .label {
+    font-size: 1.5rem;
+  }
   .withdrawal {
     position: relative;
     height: 100%;
@@ -40,6 +46,7 @@ const StyleProfileContainer = styled.div`
     align-items: center;
   }
   .tagGap {
+    font-size: 1.6rem;
     gap: 1rem;
   }
   .imgContainer {
@@ -47,6 +54,7 @@ const StyleProfileContainer = styled.div`
     position: relative;
     .userImg {
       width: 100%;
+      height: 100%;
       border-radius: 10px;
       &:hover {
         filter: brightness(0.8);
@@ -75,19 +83,19 @@ const StyleProfileContainer = styled.div`
   }
 
   .infoContainer {
-    width: 60%;
+    width: 100%;
     border: 1px solid var(--black-100);
     border-radius: 10px;
-    padding: 2.5rem;
+    padding: 2rem;
     justify-content: space-between;
     gap: 2rem;
     position: relative;
 
     .editProfile {
       position: absolute;
-      top: 2.5rem;
-      right: 2.5rem;
-      gap: 1rem;
+      top: 2rem;
+      right: 2rem;
+      gap: 0.5rem;
       display: flex;
       svg {
         cursor: pointer;
@@ -104,7 +112,7 @@ const StyleProfileContainer = styled.div`
     }
 
     .gap {
-      gap: 3rem;
+      gap: 1rem;
     }
     .infoInner {
       position: relative;
@@ -116,11 +124,8 @@ const StyleProfileContainer = styled.div`
 `;
 
 const EditTagContainer = styled.div`
-  position: absolute;
-  bottom: 0;
-  right: 0;
+  margin-top: 2rem;
   gap: 2rem;
-
   p {
     cursor: pointer;
     color: var(--error);
@@ -150,12 +155,13 @@ const Tag = styled.div`
 `;
 
 export default function ProfileCard({ id, data }) {
-  const { toAbout } = useNav();
+  const { toAbout, toProfile } = useNav();
   const [isEdit, setIsEdit] = useState({ profile: false, password: false, withDrawal: false });
   const [profile, setProfile] = useState({
     email: data.email,
     userName: data.userName,
     userImg: data.userImg,
+    isWorking: data.isWorking,
     age: data.age,
     tags: data.tags,
     aboutMe: data.aboutMe,
@@ -177,14 +183,27 @@ export default function ProfileCard({ id, data }) {
       value: profile.tags,
       curString: '',
     },
+    isWorking: {
+      value: profile.isWorking,
+    },
   });
   const [editPassword, setEditPassword] = useState({
-    prevPassword: '',
-    newPassword: '',
-    newPassword2: '',
+    prevPassword: {
+      value: '',
+      error: '',
+    },
+    newPassword: {
+      value: '',
+      error: '',
+    },
+    newPassword2: {
+      value: '',
+      error: '',
+    },
   });
   const { userId } = useParams();
   const dispatch = useDispatch();
+  const user = useSelector((state) => state.user);
 
   const handleTagKeyDown = (e) => {
     if (e.code !== 'Enter' && e.code !== 'NumpadEnter') return;
@@ -203,25 +222,70 @@ export default function ProfileCard({ id, data }) {
   useEffect(() => {}, []);
 
   const handleEditProfile = () => {
-    console.log('수정 요청 함수 실행');
-    const responseBody = {
-      aboutMe: editProfile.aboutMe.value,
-      userName: editProfile.userName.value,
-      age: editProfile.age.value,
-      tags: editProfile.tags.value,
-    };
-    console.log(responseBody);
+    console.log('프로필 수정 요청');
+    try {
+      const responseBody = {
+        aboutMe: editProfile.aboutMe.value,
+        userName: editProfile.userName.value,
+        age: editProfile.age.value,
+        tags: editProfile.tags.value,
+        isWorking: editProfile.isWorking.value,
+      };
+      api.patch(`members/${userId}`, responseBody).then(() => {
+        console.log('프로필 수정 성공');
+        toProfile(userId);
+      });
+    } catch (error) {
+      console.log(error);
+    }
   };
 
   const handleEditPassword = async (e) => {
     e.preventDefault();
+    console.log('비밀번호 수정 요청');
     try {
-      api
-        .patch(`/members/password/${userId}`, {
-          prevPassword: editPassword.prevPassword,
-          newPassword: editPassword.newPassword,
-        })
-        .then((el) => console.log(el));
+      if (
+        editPassword.newPassword === editPassword.newPassword2 &&
+        isValidPassword(editPassword.prevPassword) &&
+        isValidPassword(editPassword.newPassword) &&
+        isValidPassword(editPassword.newPassword2)
+      ) {
+        setEditPassword({
+          ...editPassword,
+          prevPassword: { ...editPassword.prevPassword, error: '' },
+          newPassword: { ...editPassword.newPassword, error: '' },
+          newPassword2: { ...editPassword.newPassword, error: '' },
+        });
+        api
+          .patch(`/members/password/${userId}`, {
+            prevPassword: editPassword.prevPassword,
+            newPassword: editPassword.newPassword,
+          })
+          .then((el) => {
+            alert('비밀번호 변경이 완료되었습니다');
+          });
+      } else if (!(editPassword.newPassword === editPassword.newPassword2)) {
+        setEditPassword({
+          ...editPassword,
+          newPassword: { ...editPassword.newPassword, error: '새 비밀번호는 같아야 합니다.' },
+          newPassword2: { ...editPassword.newPassword, error: '새 비밀번호는 같아야 합니다.' },
+        });
+      } else if (!isValidPassword(editPassword.prevPassword)) {
+        setEditPassword({
+          ...editPassword,
+          prevPassword: { ...editPassword.prevPassword, error: '다시 입력해주세요' },
+        });
+      } else if (!isValidPassword(editPassword.newPassword)) {
+        setEditPassword({
+          ...editPassword,
+          newPassword: { ...editPassword.newPassword, error: '다시 입력해주세요' },
+        });
+      } else if (isValidPassword(editPassword.newPassword)) {
+        setEditPassword({
+          ...editPassword,
+          newPassword2: { ...editPassword.newPassword2, error: '다시 입력해주세요' },
+        });
+      }
     } catch (error) {
       console.log(error);
     }
@@ -229,22 +293,52 @@ export default function ProfileCard({ id, data }) {
 
   const handleClickWithdrawal = () => {
     console.log('회원탈퇴 요청');
-    api.delete(`/members/${userId}`).then((el) => console.log(el));
-    dispatch(deleteUser());
-    toAbout();
+    if (window.confirm('정말 탈퇴하시겠습니까 ?')) {
+      api.delete(`/members/${userId}`).then((el) => console.log(el));
+      dispatch(deleteUser());
+      alert('이용해주셔서 감사합니다.');
+      toAbout();
+    }
   };
 
-  const handleClickUserImgUpload = () => {
-    console.log('회원 이미지 업로드');
+  const fileInputRef = React.createRef();
+
+  const handleClickUserImg = () => {
+    fileInputRef.current.click();
+  };
+
+  const handleFileChange = (e) => {
+    console.log('유저 이미지 교체 요청');
+    const file = e.target.files[0];
+    api.patch(`/members/profileImg/${userId}`).then((el) => {
+      setProfile({ ...profile, userImg: el.data.imgUrl });
+      dispatch(updateUser({ userInfo: { ...user.userInfo, imgUrl: el.data.imgUrl } }));
+    });
+    console.log(file);
   };
 
   return (
     <StyleProfileContainer id={id} className="row">
       <div className="imgContainer">
-        <img className="userImg" src={profile.userImg} alt="userImage" />
-        <div className="editImg" onClick={handleClickUserImgUpload}>
-          <PiUploadSimple color="black" size="30" />
-        </div>
+        <img
+          className="userImg"
+          src={profile.userImg ? profile.userImg : userDefaultImg}
+          alt="userImage"
+        />
+        {user.isLogin && Number(userId) === user.userInfo.memberId && (
+          <>
+            <input
+              type="file"
+              accept=".png, .jpg, .jpeg"
+              className="hidden"
+              onChange={handleFileChange}
+              ref={fileInputRef}
+            />
+            <div className="editImg" onClick={handleClickUserImg}>
+              <PiUploadSimple color="black" size="30" />
+            </div>
+          </>
+        )}
       </div>
       <div className="infoContainer col">
         {!isEdit.profile && !isEdit.password && !isEdit.withDrawal && (
@@ -256,7 +350,7 @@ export default function ProfileCard({ id, data }) {
             EditTagContainer={EditTagContainer}
           />
         )}
-        {isEdit.profile && (
+        {isEdit.profile && user.isLogin && Number(userId) === user.userInfo.memberId && (
           <EditProfile
             editProfile={editProfile}
             setEditProfile={setEditProfile}
@@ -267,7 +361,7 @@ export default function ProfileCard({ id, data }) {
             Tag={Tag}
           />
         )}
-        {isEdit.password && (
+        {isEdit.password && user.isLogin && Number(userId) === user.userInfo.memberId && (
           <EditPassword
             isEdit={isEdit}
             setIsEdit={setIsEdit}
@@ -276,7 +370,7 @@ export default function ProfileCard({ id, data }) {
             handleEditPassword={handleEditPassword}
           />
         )}
-        {isEdit.withDrawal && (
+        {isEdit.withDrawal && user.isLogin && Number(userId) === user.userInfo.memberId && (
           <Withdrawal
             handleClickWithdrawal={handleClickWithdrawal}
             isEdit={isEdit}
